Move filter spec validation into its own module
Relocate validate_filter_spec (and its helpers and the query-tree limit
constants) out of endpoint_capability into a new filter_spec module, and move
the corresponding tests into test_filter_spec. endpoint_capability keeps only
the capability metadata; filter_spec imports COMPONENT_INPUT_SCHEMAS from it.

Pure move, no behavior change.
